+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Glycol-water mixtures can now be selected from within the GUI (issue #174).
 - Pygfunction media object can be imported into the FluidData object in GHEtool (issue #174).
 - Temperature database (Europe) based on data from (Xing, 2014) (issue #178).
+- Yearly heating/cooling load in LoadClass (issue #180).
 
 ## Changed
 - GUI was moved to a separate project: [ScenarioGUI](https://github.com/tblanke/ScenarioGUI).